+++ cognite/client/_api/data_modeling/instances.py
@@ -37,6 +37,10 @@ from cognite.client.data_classes.data_modeling.instances import (
     NodeApplyResultList,
     NodeList,
 )
+from cognite.client.data_classes.data_modeling.query import (
+    Query,
+    QueryResult,
+)
 from cognite.client.data_classes.data_modeling.views import View
 from cognite.client.utils._identifier import DataModelingIdentifierSequence
 
@@ -732,6 +736,97 @@ class InstancesAPI(APIClient):
             return HistogramValue.load(res.json()["items"][0]["aggregates"][0])
         else:
             return [HistogramValue.load(item["aggregates"][0]) for item in res.json()["items"]]
+
+    def query(self, query: Query) -> QueryResult:
+        """`Advanced query interface for nodes/edges. <https://developer.cognite.com/api/v1/#tag/Instances/operation/queryContent>`_
+
+        The Data Modelling API exposes an advanced query interface. The query interface supports parameterization,
+        recursive edge traversal, chaining of result sets, and granular property selection.
+
+        Args:
+            query: Query.
+
+        Returns:
+            QueryResult: The resulting nodes and/or edges from the query.
+
+        Examples:
+
+            Find actors in movies released before 2000 sorted by actor name:
+
+                >>> from cognite.client import CogniteClient
+                >>> from cognite.client.data_classes.data_modeling.query import Query, Select, NodeResultSetExpression, EdgeResultSetExpression, SourceSelector
+                >>> from cognite.client.data_classes.data_modeling.filters import Range, Equals
+                >>> from cognite.client.data_classes.data_modeling.ids import ViewId
+                >>> c = CogniteClient()
+                >>> movie_id = ViewId("mySpace", "MovieView", "v1")
+                >>> actor_id = ViewId("mySpace", "ActorView", "v1")
+                >>> query = Query(
+                ...         with_ = {
+                ...             "movies": NodeResultSetExpression(filter=Range(movie_id.as_property_ref("releaseYear"), lt=2000)),
+                ...             "actors_in_movie": EdgeResultSetExpression(from_="movies", filter=Equals(["edge", "type"], {"space": movie_id.space, "externalId": "Movie.actors"})),
+                ...             "actors": NodeResultSetExpression(from_="actors_in_movie"),
+                ...         },
+                ...         select = {
+                ...             "actors": Select(
+                ...                            [SourceSelector(actor_id, ["name"])], sort=[actor_id.as_property_ref("name")]),
+                ...         },
+                ... )
+                >>> res = c.data_modeling.instances.query(query)
+        """
+        return self._query_or_sync(query, "query")
+
+    def sync(self, query: Query) -> QueryResult:
+        """`Subscription to changes for nodes/edges. <https://developer.cognite.com/api/v1/#tag/Instances/operation/syncContent>`_
+
+        Subscribe to changes for nodes and edges in a project, matching a supplied filter.
+
+        Args:
+            query: Query.
+
+        Returns:
+            QueryResult: The resulting nodes and/or edges from the query.
+
+        Examples:
+
+            Find actors in movies released before 2000 sorted by actor name:
+
+                >>> from cognite.client import CogniteClient
+                >>> from cognite.client.data_classes.data_modeling.query import Query, Select, NodeResultSetExpression, EdgeResultSetExpression, SourceSelector
+                >>> from cognite.client.data_classes.data_modeling.filters import Range, Equals
+                >>> from cognite.client.data_classes.data_modeling.ids import ViewId
+                >>> c = CogniteClient()
+                >>> movie_id = ViewId("mySpace", "MovieView", "v1")
+                >>> actor_id = ViewId("mySpace", "ActorView", "v1")
+                >>> query = Query(
+                ...         with_ = {
+                ...             "movies": NodeResultSetExpression(filter=Range(movie_id.as_property_ref("releaseYear"), lt=2000)),
+                ...             "actors_in_movie": EdgeResultSetExpression(from_="movies", filter=Equals(["edge", "type"], {"space": movie_id.space, "externalId": "Movie.actors"})),
+                ...             "actors": NodeResultSetExpression(from_="actors_in_movie"),
+                ...         },
+                ...         select = {
+                ...             "actors": Select(
+                ...                            [SourceSelector(actor_id, ["name"])], sort=[actor_id.as_property_ref("name")]),
+                ...         },
+                ... )
+                >>> res = c.data_modeling.instances.sync(query)
+                >>> # Added a new movie with actors released before 2000
+                >>> query.cursors = res.cursors
+                >>> res_new = c.data_modeling.instances.sync(query)
+
+            In the last example, the res_new will only contain the actors that have been added with the new movie.
+        """
+        return self._query_or_sync(query, "sync")
+
+    def _query_or_sync(self, query: Query, endpoint: Literal["query", "sync"]) -> QueryResult:
+        body = query.dump(camel_case=True)
+
+        result = self._post(url_path=self._RESOURCE_PATH + f"/{endpoint}", json=body)
+
+        json_payload = result.json()
+        default_by_reference = query.instance_type_by_result_expression()
+        results = QueryResult.load(json_payload["items"], default_by_reference, json_payload["nextCursor"])
+
+        return results
 
     @overload
     def list(